# frontend/src/backend/openapi-schema.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/projects":{"get":{"summary":"Read Projects","operationId":"read_projects_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Projects Projects Get","type":"array","items":{"$ref":"#/components/schemas/Project"}}}}}}}},"/project/{project_id}":{"get":{"summary":"Read Project","operationId":"read_project_project__project_id__get","parameters":[{"required":true,"schema":{"title":"Project Id","type":"integer"},"name":"project_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/project/{project_id}/create_review":{"post":{"summary":"Create Review","operationId":"create_review_project__project_id__create_review_post","parameters":[{"required":true,"schema":{"title":"Project Id","type":"integer"},"name":"project_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Project":{"title":"Project","required":["id","name","slug","reviews"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"},"reviews":{"title":"Reviews","type":"array","items":{"$ref":"#/components/schemas/Review"}}}},"Review":{"title":"Review","required":["id","text","score","user"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"text":{"title":"Text","type":"string"},"score":{"title":"Score","type":"integer"},"user":{"$ref":"#/components/schemas/User"}}},"ReviewCreate":{"title":"ReviewCreate","required":["score","text","address"],"type":"object","properties":{"score":{"title":"Score","type":"integer"},"text":{"title":"Text","type":"string"},"address":{"title":"Address","type":"string"}}},"User":{"title":"User","required":["id","address"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"address":{"title":"Address","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}